server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==9.5.0
numpy==2.0.2
onnxruntime==1.22.1
rembg==2.0.55
torch
requests==2.32.3

